feat: Task 033 Student#4: sponsor data model
I have a couple questions about this implementation. The attribute benefits it is defined as a list but by the time it has this constraint "shorter than 101 characters" as a general benefits I saw it as a list written as a long string with the benefits. The other question is about the definition of sponsor which said that it was a role so I defined coming from Abstract role as the other roles.